changed retrun type of mathScore method in Searchable interface from int to double because string similarity library work with double type!

+++ src/com/Searchable.java
@@ -1,5 +1,5 @@
 package com;
 
 public interface Searchable {
-    int matchScore(String string);
+    double matchScore(String string);
 }
